+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [PR #2106](https://github.com/rqlite/rqlite/pull/2106), [PR #2126](https://github.com/rqlite/rqlite/pull/2126): Add initial CDC Service implementation.
 - [PR #2128](https://github.com/rqlite/rqlite/pull/2128): Support CDC highwatermarking control.
 - [PR #2129](https://github.com/rqlite/rqlite/pull/2129): CDC Service improvements and refactoring.
+- [PR #2131](https://github.com/rqlite/rqlite/pull/2131): FIFO queue for CDC service.
 
 ## v8.38.2 (June 23rd 2025)
 ### Implementation changes and bug fixes